feat: refactored file interface

+++ include/kernel/fs.h
@@ -41,7 +41,7 @@ ssize_t fs_readv(int fd, const struct iovec *iov, int iovcnt);
 ssize_t fs_writev(int fd, const struct iovec *iov, int iovcnt);
 ssize_t fs_readdir(int fd, void *dirp, size_t len);
 off_t fs_lseek(int fd, off_t offset, int whence);
-int fs_ioctl(int fd, unsigned long request, void *argp);
+int fs_ioctl(int fd, unsigned int request, void *argp);
 int fs_fcntl(int fd, int cmd, unsigned long arg);
 int fs_ftruncate(int fd, off_t length);
 int fs_fstat(int fd, struct stat *stat);
